requirements.txt
discord.py
feedparser
flask
